Migrate check_second_toc_level to BookTest
