+++ tests/testsuite/toc/basic_toc/book.toml
@@ -1,0 +1,2 @@
+[book]
+title = "basic_toc"
