+++ tests/testsuite/toc/basic_toc/src/SUMMARY.md
@@ -1,0 +1,23 @@
+# Summary
+
+[Prefix 1](prefix1.md)
+[Prefix 2](prefix2.md)
+
+- [With Readme](README.md)
+    - [Nested Index](nested/index.md)
+    - [Nested two](nested/two.md)
+- [Draft]()
+
+---
+
+# Deep Nest
+
+- [Deep Nest 1](deep/index.md)
+    - [Deep Nest 2](deep/a/index.md)
+        - [Deep Nest 3](deep/a/b/index.md)
+            [Deep Nest 4](deep/a/b/c/index.md)
+
+---
+
+[Suffix 1](suffix1.md)
+[Suffix 2](suffix2.md)
