+++ tests/testsuite/toc/basic_toc/src/nested/index.md
@@ -1,0 +1,1 @@
+# Nested Index
