+++ tests/testsuite/toc/basic_toc/src/prefix1.md
@@ -1,0 +1,1 @@
+# Prefix 1
