+++ tests/testsuite/toc/basic_toc/src/prefix2.md
@@ -1,0 +1,1 @@
+# Prefix 2
