+++ tests/testsuite/toc/basic_toc/src/suffix1.md
@@ -1,0 +1,1 @@
+# Suffix 1
